+++ src/main/java/com/soundcloud/prometheus/hystrix/HystrixPrometheusMetricsPublisherThreadPool.java
@@ -30,7 +30,6 @@ public class HystrixPrometheusMetricsPublisherThreadPool implements HystrixMetri
 
     private final SortedMap<String, String> labels;
     private final boolean exportProperties;
-    private final boolean exportDeprecatedMetrics;
 
     private final HystrixThreadPoolMetrics metrics;
     private final HystrixThreadPoolProperties properties;
@@ -39,7 +38,7 @@ public class HystrixPrometheusMetricsPublisherThreadPool implements HystrixMetri
 
     public HystrixPrometheusMetricsPublisherThreadPool(
             HystrixMetricsCollector collector, HystrixThreadPoolKey key, HystrixThreadPoolMetrics metrics,
-            HystrixThreadPoolProperties properties, boolean exportProperties, boolean exportDeprecatedMetrics,
+            HystrixThreadPoolProperties properties, boolean exportProperties,
             HystrixMetricsPublisherThreadPool delegate) {
 
         this.metrics = metrics;
@@ -48,7 +47,6 @@ public class HystrixPrometheusMetricsPublisherThreadPool implements HystrixMetri
         this.exportProperties = exportProperties;
         this.labels = new TreeMap<String, String>();
         this.labels.put("pool_name", key.name());
-        this.exportDeprecatedMetrics = exportDeprecatedMetrics;
         this.delegate = delegate;
     }
 
@@ -59,75 +57,47 @@ public class HystrixPrometheusMetricsPublisherThreadPool implements HystrixMetri
         String currentStateDoc = "Current state of thread-pool partitioned by pool_name.";
         addGauge("thread_active_count", currentStateDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getCurrentActiveCount();
             }
         });
         addGauge("completed_task_count", currentStateDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getCurrentCompletedTaskCount();
             }
         });
         addGauge("largest_pool_size", currentStateDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getCurrentLargestPoolSize();
             }
         });
         addGauge("total_task_count", currentStateDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getCurrentTaskCount();
             }
         });
         addGauge("queue_size", currentStateDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getCurrentQueueSize();
             }
         });
 
         String rollDoc = "Rolling count partitioned by pool_name.";
-        if (exportDeprecatedMetrics) {
-            addGauge("rolling_max_active_threads", "DEPRECATED, use rolling_active_threads_max instead: " + rollDoc,
-                    new Callable<Number>() {
-                        @Override
-                        public Number call() throws Exception {
-                            return metrics.getRollingMaxActiveThreads();
-                        }
-                    });
-        }
         addGauge("rolling_active_threads_max", rollDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getRollingMaxActiveThreads();
             }
         });
 
-        if (exportDeprecatedMetrics) {
-            addGauge("rolling_count_threads_executed", "DEPRECATED, use rate(threads_executed_total) instead: " + rollDoc,
-                    new Callable<Number>() {
-                        @Override
-                        public Number call() throws Exception {
-                            return metrics.getRollingCountThreadsExecuted();
-                        }
-                    });
-        }
-
         String totalDoc = "Cumulative count partitioned by pool_name.";
-        if (exportDeprecatedMetrics) {
-            addGauge("count_threads_executed", "DEPRECATED: " + totalDoc,
-                    new Callable<Number>() {
-                        @Override
-                        public Number call() throws Exception {
-                            return metrics.getCumulativeCountThreadsExecuted();
-                        }
-                    });
-        }
         addGauge("threads_executed_total", totalDoc, new Callable<Number>() {
             @Override
-            public Number call() throws Exception {
+            public Number call() {
                 return metrics.getCumulativeCountThreadsExecuted();
             }
         });
@@ -136,25 +106,25 @@ public class HystrixPrometheusMetricsPublisherThreadPool implements HystrixMetri
             String doc = "Configuration property partitioned by pool_name.";
             addGauge("property_value_core_pool_size", doc, new Callable<Number>() {
                 @Override
-                public Number call() throws Exception {
+                public Number call() {
                     return properties.coreSize().get();
                 }
             });
             addGauge("property_value_keep_alive_time_in_minutes", doc, new Callable<Number>() {
                 @Override
-                public Number call() throws Exception {
+                public Number call() {
                     return properties.keepAliveTimeMinutes().get();
                 }
             });
             addGauge("property_value_queue_size_rejection_threshold", doc, new Callable<Number>() {
                 @Override
-                public Number call() throws Exception {
+                public Number call() {
                     return properties.queueSizeRejectionThreshold().get();
                 }
             });
             addGauge("property_value_max_queue_size", doc, new Callable<Number>() {
                 @Override
-                public Number call() throws Exception {
+                public Number call() {
                     return properties.maxQueueSize().get();
                 }
             });